summary view improvements

+++ Dotnet/Pages/OrderSummary.cshtml.cs
@@ -1,32 +1,77 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 using Shop.BusinessObjects;
 using Shop.Pages.Shared;
 using Shop.Services;
 
 namespace Shop.Pages
 {
-    public class OrderSummaryModel: LayoutModel
+    public class OrderSummaryModel : LayoutModel
     {
-        public ShipmentInfo ShipmentInfo { get; set; }
+        [BindProperty] public ShipmentInfo ShipmentInfo { get; set; }
+        [BindProperty] public BasketSet BasketSet { get; set; }
         private IOrderService _orderService;
         private IBasketService _basketService;
-        public OrderSummaryModel(ILayoutService layoutService, IOrderService orderService, IBasketService basketService) : base(layoutService)
+        public List<float> price;
+
+        public OrderSummaryModel(ILayoutService layoutService, IOrderService orderService, IBasketService basketService)
+            : base(layoutService)
         {
             _orderService = orderService;
             _basketService = basketService;
         }
 
-        public void OnGet()
+        public IActionResult OnGet()
         {
-            
+            switch (_orderService.CurrentOrderState())
+            {
+                case OrderState.Basket:
+                    return RedirectToPage("Basket");
+                case OrderState.Shipment:
+                    return RedirectToPage("ShipmentInfo");
+                case OrderState.Summary:
+                    break;
+                default:
+                    _orderService.SetCurrentOrderState(OrderState.Basket);
+                    return RedirectToPage("Basket");
+            }
+
+            BasketSet = _basketService.GetOrderItems();
             ShipmentInfo = _orderService.GetShipmentInfoFromSession() ?? new ShipmentInfo();
-            _basketService.RemoveOrderedItemsFromCookie();
+
             initializeLayout();
+
+            return Page();
         }
 
-        public void OnGetSingleOrder(ShipmentInfo shipmentInfo)
+        public async Task<IActionResult> OnPostAsync()
         {
-            initializeLayout();
-            ShipmentInfo = shipmentInfo;   
+            if (!ModelState.IsValid)
+            {
+                initializeLayout();
+                return Page();
+            }
+
+            _basketService.RemoveOrderedItemsFromCookie();
+
+            return RedirectToPage("Orders");
+//            return RedirectToPage("OrderSummary", "SingleOrder", ShipmentInfo);
+        }
+
+//        public void OnGetSingleOrder(ShipmentInfo shipmentInfo)
+//        {
+//            initializeLayout();
+//            ShipmentInfo = shipmentInfo;
+//        }
+        public IActionResult OnPostBack()
+        {
+            return RedirectToPage("ShipmentInfo");
+        }
+
+        public IActionResult OnPostCancel()
+        {
+            return RedirectToPage("Basket");
         }
     }
 }
